[Fix] Build errors
